# On common URLs, this should speed things up a bit in parse_host.