--- a/src/unicode.cpp
+++ b/src/unicode.cpp
@@ -24,6 +24,31 @@
 
 namespace ada::unicode {
 
+  constexpr bool to_lower_ascii(char * input, size_t length) noexcept {
+    auto broadcast = [](uint8_t v) -> uint64_t { return 0x101010101010101 * v; };
+    uint64_t broadcast_80 = broadcast(0x80);
+    uint64_t broadcast_Ap = broadcast(128 - 'A');
+    uint64_t broadcast_Zp = broadcast(128 - 'Z');
+    uint8_t non_ascii = 0;
+    size_t i = 0;
+
+    for (; i + 7 < length; i += 8) {
+      uint64_t word{};
+      memcpy(&word, input + i, sizeof(word));
+      non_ascii |= (word & broadcast_80);
+      word ^= (((word+broadcast_Ap)^(word+broadcast_Zp))&broadcast_80)>>2;
+      memcpy(input + i, &word, sizeof(word));
+    }
+    if (i < length) {
+      uint64_t word{};
+      memcpy(&word, input + i, length - i);
+      non_ascii |= (word & broadcast_80);
+      word ^= (((word+broadcast_Ap)^(word+broadcast_Zp))&broadcast_80)>>2;
+      memcpy(input + i, &word, length - i);
+    }
+    return non_ascii == 0;
+  }
+
   ada_really_inline constexpr bool has_tabs_or_newline(std::string_view user_input) noexcept {
     auto has_zero_byte = [](uint64_t v) {
       return ((v - 0x0101010101010101) & ~(v)&0x8080808080808080);