--- a/tests/wpt/ada_extra_urltestdata.json
+++ b/tests/wpt/ada_extra_urltestdata.json
@@ -166,6 +166,11 @@
     "base": "about:blank",
     "failure": true
   },
+  {
+    "input": "schéme://example.com",
+    "base": "about:blank",
+    "failure": true
+  },
   "# Credentials",
   {
     "input": "scheme://username@@@@example.com",