Before... (bbc_bench)

```
----------------------------------------------------------------------------
Benchmark                  Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------
BasicBench_AdaURL       2872 ns         2872 ns       241453 speed=300.508M/s time/byte=3.3277ns time/url=319.089ns url/s=3.13392M/s
BasicBench_CURL         6501 ns         6501 ns       109582 speed=132.757M/s time/byte=7.53256ns time/url=722.289ns url/s=1.38449M/s
```

After... (bbc_bench)

```
----------------------------------------------------------------------------
Benchmark                  Time             CPU   Iterations UserCounters...
----------------------------------------------------------------------------
BasicBench_AdaURL       2508 ns         2508 ns       269107 speed=344.136M/s time/byte=2.90583ns time/url=278.637ns url/s=3.5889M/s
BasicBench_CURL         6486 ns         6486 ns       108238 speed=133.051M/s time/byte=7.51589ns time/url=720.691ns url/s=1.38756M/s
```